feat: create and start indexers

+++ migrations/2023-09-18-115821_create_indexers/up.sql
@@ -4,5 +4,5 @@ CREATE TABLE indexers
     id         uuid PRIMARY KEY DEFAULT uuid_generate_v4(),
     status     VARCHAR NOT NULL,
     type       VARCHAR NOT NULL,
-    process_id INT
+    process_id BIGINT
 )
